Tercera paractica integradora

+++ src/controllers/products.controller.js
@@ -6,7 +6,7 @@ import { generateProduct } from '../utils/facker.js';
 import { logger } from '../utils/logger.js';
 
 class ProductController {
-  async getAllProducts(req, res, next) {
+  async getAllProducts(_, res, next) {
     try {
       const product = await productService.getAllProducts();
       if (!product) {
@@ -50,11 +50,13 @@ class ProductController {
     try {
       const { title, description, code, price, stock, category, thumbnail } = req.body;
 
+      const owner = req.session.user.email;
+
       if (!title || !description || !code || !price || !stock || !category || !thumbnail) {
         throw new CustomError(EErrors.INVALID_INPUT_ERROR.code, EErrors.INVALID_INPUT_ERROR.name, EErrors.INVALID_INPUT_ERROR.cause, EErrors.INVALID_INPUT_ERROR.message);
       }
 
-      let product = new ProductDTO({ title, description, code, price, stock, category, thumbnail });
+      let product = new ProductDTO({ title, description, code, price, stock, category, thumbnail, owner });
       const productCreated = await productService.create(product);
 
       return res.status(201).json({
@@ -95,6 +97,15 @@ class ProductController {
     try {
       const { id } = req.params;
       logger.debug(`Product id received by parameter: ${id}`);
+
+      const { user } = req.user;
+
+      const productFound = await this.getProductById(id);
+
+      if (productFound.rol === 'premium' && productFound.owner != user.email) {
+        throw new CustomError(EErrors.PRODUCT_OWNER_DELETE.code, EErrors.PRODUCT_OWNER_DELETE.name, EErrors.PRODUCT_OWNER_DELETE.cause, EErrors.PRODUCT_OWNER_DELETE.message);
+      }
+
       const productDeleted = await productService.deleteOne(id);
       return res.status(200).json({
         status: 'success',
@@ -107,7 +118,7 @@ class ProductController {
     }
   }
 
-  async mockingProducts(req, res) {
+  async mockingProducts(_, res) {
     const products = [];
     for (let i = 0; i < 100; i++) {
       products.push(generateProduct());
